+++ lib/presentation/pages/auth/widgets/auth_white_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:todo_app/presentation/core/theme/todo_card_color.dart';
+import 'package:todo_app/presentation/core/theme/todo_theme.dart';
 import 'package:todo_app/presentation/providers/auth/auth_provider.dart';
 
 class AuthWhiteButton extends ConsumerWidget {
